test: refactor unit test to create user andate user. create test to update password

+++ src/main/java/com/anthony/calorie_counter/service/impl/UserService.java
@@ -4,9 +4,9 @@ import com.anthony.calorie_counter.entity.RoleModel;
 import com.anthony.calorie_counter.entity.UserModel;
 import com.anthony.calorie_counter.enums.UserRole;
 import com.anthony.calorie_counter.exceptions.EntityDataNotFoundException;
-import com.anthony.calorie_counter.repository.RoleRepository;
 import com.anthony.calorie_counter.repository.UserRepository;
 import com.anthony.calorie_counter.service.IUserService;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -20,12 +20,10 @@ import java.util.UUID;
 @Service
 public class UserService implements IUserService, UserDetailsService {
     private final UserRepository userRepository;
-    private final RoleRepository roleRepository;
     private final PasswordEncoder passwordEncoder;
 
-    public UserService(UserRepository userRepository, RoleRepository roleRepository, PasswordEncoder passwordEncoder) {
+    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder) {
         this.userRepository = userRepository;
-        this.roleRepository = roleRepository;
         this.passwordEncoder = passwordEncoder;
     }
 
@@ -37,27 +35,40 @@ public class UserService implements IUserService, UserDetailsService {
 
     @Override
     public UserModel create(UserModel userModel) {
-        userModel.addRole(findRoleById(UserRole.ROLE_USER.getRole()));
+        userModel.addRole(new RoleModel(UserRole.ROLE_USER));
         userModel.setPassword(passwordEncoder.encode(userModel.getPassword()));
-        return save(userModel);
+        return userRepository.save(userModel);
     }
 
     @Override
     public UserModel updateUser(UUID id, UserModel newUserModelData) {
-        UserModel userModel = findById(id);
-        userModel.setName(newUserModelData.getName());
-        userModel.setEmail(newUserModelData.getEmail());
-        userModel.setPhoneNumber(newUserModelData.getPhoneNumber());
-        return save(userModel);
+        try {
+            UserModel userModel = userRepository.getReferenceById(id);
+            userModel.setName(newUserModelData.getName());
+            userModel.setEmail(newUserModelData.getEmail());
+            userModel.setPhoneNumber(newUserModelData.getPhoneNumber());
+            return userRepository.save(userModel);
+        } catch (EntityNotFoundException e) {
+            throw new EntityDataNotFoundException("User not found with id: " + id);
+        }
     }
 
     @Transactional
     public void updatePassword(UUID id, String newPassword) {
-        userRepository.updatePasswordByUserId(id, passwordEncoder.encode(newPassword));
+        try {
+            UserModel userModel = userRepository.getReferenceById(id);
+            userModel.setPassword(passwordEncoder.encode(newPassword));
+            userRepository.save(userModel);
+        } catch (EntityNotFoundException e) {
+            throw new EntityDataNotFoundException("User not found with id: " + id);
+        }
     }
 
     @Override
     public void deleteById(UUID id) {
+        if (!userRepository.existsById(id)) {
+            throw new EntityDataNotFoundException("User not found with id: " + id);
+        }
         userRepository.deleteById(id);
     }
 
@@ -66,28 +77,23 @@ public class UserService implements IUserService, UserDetailsService {
         return userRepository.findAll(pageable);
     }
 
-    @Transactional(readOnly = true)
-    public RoleModel findRoleById(Long id) {
-        return roleRepository.findById(id)
-                .orElseThrow(() -> new EntityDataNotFoundException("Role not found with id: " + id));
-    }
-
-    @Transactional
-    private UserModel save(UserModel user) {
-        return userRepository.save(user);
-    }
+//    @Transactional(readOnly = true)
+//    public RoleModel findRoleById(Long id) {
+//        return roleRepository.findById(id)
+//                .orElseThrow(() -> new EntityDataNotFoundException("Role not found with id: " + id));
+//    }
 
     @Override @Transactional
     public UserModel promoteToAdmin(UUID id) {
         UserModel userModel = findById(id);
-        userModel.addRole(findRoleById(UserRole.ROLE_ADMIN.getRole()));
+        userModel.addRole(new RoleModel(UserRole.ROLE_ADMIN));
         return userRepository.save(userModel);
     }
 
     @Override @Transactional
     public UserModel demoteFromAdmin(UUID id) {
         UserModel userModel = findById(id);
-        userModel.removeRole(UserRole.ROLE_ADMIN.getRole());
+        userModel.removeRoleById(UserRole.ROLE_ADMIN.getRole());
         return userRepository.save(userModel);
     }
